[refactor] Structure React components
