+++ src/components/plantFacts.jsx
